Refactor examples (#110)
* Separate MatPower from other test utilities.

* Update figure path in power flow example README file.

* Fix warning coming from Variable class.

* Fix warnings in Matpower tests.

* Reduce output from phasor dynamics Example 1

* Minor fixes to DG and microgrid examples.


---------

+++ src/Utilities/FileIO.hpp
@@ -189,7 +189,7 @@ namespace
     if (std::regex_match(line, matches, pat))
     {
       std::string s = matches[1];
-      mp.baseMVA    = std::atoi(s.c_str());
+      mp.baseMVA    = std::atof(s.c_str());
     }
     else
     {
